fix(pol): refresh boosts after tx, renaming components
