+++ apps/hub/src/app/validators/validator/unboost-modal.tsx
@@ -1,28 +1,38 @@
 import { useState } from "react";
+import { useUserBoostsOnValidator } from "@bera/berajs";
 import { Button } from "@bera/ui/button";
 import { Dialog, DialogContent } from "@bera/ui/dialog";
+
+import { UnboostContent } from "./unboost-content";
+import { ApiValidatorFragment } from "@bera/graphql/pol/api";
 import { Address } from "viem";
 
-import { DelegateContent } from "./delegate-content";
-import { ActionButton } from "@bera/shared-ui";
-
-export const DelegateModal = ({
+export const UnbondModal = ({
   validator,
   setIsValidatorDataLoading,
 }: {
-  validator: Address;
+  validator: ApiValidatorFragment | undefined;
   setIsValidatorDataLoading: (loading: boolean) => void;
 }) => {
   const [open, setOpen] = useState(false);
 
+  const { data: userBoosts } = useUserBoostsOnValidator({
+    pubkey: validator?.pubkey as Address,
+  });
+
+  if (!validator) return null;
   return (
     <>
-      <ActionButton>
-        <Button onClick={() => setOpen(true)}>Boost</Button>
-      </ActionButton>
+      <Button
+        variant="ghost"
+        onClick={() => setOpen(true)}
+        disabled={Number(userBoosts?.activeBoosts) <= 0}
+      >
+        Unboost
+      </Button>
       <Dialog open={open} onOpenChange={setOpen}>
         <DialogContent className="w-full md:w-[550px]">
-          <DelegateContent
+          <UnboostContent
             onSuccess={() => {
               setOpen(false);
             }}
